Move datasource.js and swagger.js to /aspsp-resource-server/
Move corresponding test files to /test/aspsp-resource-server/
