+++ lib/index.js
@@ -1,9 +1,9 @@
 const { Middleware } = require('swagger-express-middleware');
 const { initializeMiddleware } = require('swagger-tools');
-const { fetchSwagger } = require('./swagger.js');
+const { fetchSwagger } = require('./aspsp-resource-server/swagger');
 const { app } = require('./app.js');
 
-const dataSource = require('./datasource.js');
+const dataSource = require('./aspsp-resource-server/datasource');
 const error = require('debug')('error');
 const morgan = require('morgan');
